+++ src/utils.test.js
@@ -1,5 +1,15 @@
-import { updateResourceData } from './utils';
+import { getFullName, getPagingType, reduceUsersToMap, updateResourceData } from './utils';
 import { UNASSIGNED_FILTER_KEY, ASSIGNED_FILTER_KEY } from './constants';
+
+const MCLPagingTypes = {
+  PREV_NEXT: 'prevNext',
+  NONE: 'none',
+};
+
+jest.mock('@folio/stripes/components', () => ({
+  ...jest.requireActual('@folio/stripes/components'),
+  MCLPagingTypes,
+}));
 
 describe('updatedResourceData', () => {
   describe('when only UnAssigned filter is selected', () => {
@@ -53,5 +63,44 @@ describe('updatedResourceData', () => {
       };
       expect(updateResourceData(resourceData)).toMatchObject(expectedResourceData);
     });
+  });
+});
+
+describe('getFullName', () => {
+  it('includes preferred first name if present', () => {
+    expect(getFullName({
+      personal: {
+        firstName: 'mockFirstName',
+        preferredFirstName: 'preferred',
+      },
+    })).toEqual(', preferred (mockFirstName) ');
+  });
+
+  it('falls back to empty strings', () => {
+    expect(getFullName({
+      personal: {},
+    })).toEqual('  ');
+  });
+});
+
+describe('reduceUsersToMap', () => {
+  const mockUser = { id: 'mockId' };
+
+  it('when the users are not checked', () => {
+    expect(reduceUsersToMap([mockUser])).toEqual({ 'mockId': null });
+  });
+
+  it('when the users are checked', () => {
+    expect(reduceUsersToMap([mockUser], true)).toEqual({ 'mockId': mockUser });
+  });
+});
+
+describe('getPagingType', () => {
+  it('returns PREV_NEXT paging type', () => {
+    expect(getPagingType({})).toBe(MCLPagingTypes.PREV_NEXT);
+  });
+
+  it('returns NONE paging type', () => {
+    expect(getPagingType({ state: { uas: ['randomValue'] } })).toBe(MCLPagingTypes.NONE);
   });
 });