+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In progress
 
 * React v19: refactor away from default props for functional components. Refs UIPFU-91.
+* Migration of unit tests to Jest for ui-plugin-find-user/src /UserSearchView.js. Refs UIPFU-85.
 
 ## [7.2.0] (https://github.com/folio-org/ui-plugin-find-user/tree/v7.2.0) (2024-09-05)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v7.1.2...v7.2.0)